Put global styles in a separate stylesheet.
Putting dynamic styled component css rules into the same stylesheet
as global styles can cause global rules such as @font-face to be
recomputed.  Instead, we now create two stylesheets, one for global
styles, and one for styled-components.
