+++ src/constructors/test/injectGlobal.test.js
@@ -4,7 +4,7 @@ import expect from 'expect'
 import { shallow } from 'enzyme'
 
 import injectGlobal from '../injectGlobal'
-import styleSheet from '../../models/StyleSheet'
+import styleSheet from '../../models/GlobalStyleSheet'
 import { expectCSSMatches, resetStyled } from '../../test/utils'
 
 let styled = resetStyled()
@@ -14,7 +14,7 @@ const rule3 = 'color: blue;'
 
 describe('injectGlobal', () => {
   beforeEach(() => {
-    resetStyled()
+    resetStyled(styleSheet)
   })
 
   it(`should inject rules into the head`, () => {
@@ -45,10 +45,10 @@ describe('injectGlobal', () => {
       a {
         ${rule2}
       }
-    `)
+    `, { styleSheet })
   })
 
-  it(`should non-destructively inject styles when called after a component`, () => {
+  it(`should inject styles in a separate sheet from a component`, () => {
     const Comp = styled.div`
       ${rule3}
     `
@@ -59,14 +59,17 @@ describe('injectGlobal', () => {
         ${rule1}
       }
     `
-
+    // Test the component sheet
     expectCSSMatches(`
       .a {
         ${rule3}
       }
+    `)
+    // Test the global sheet
+    expectCSSMatches(`
       html {
         ${rule1}
       }
-    `)
+    `, { styleSheet })
   })
 });
